Conflicts:
	.classpath
	AndroidManifest.xml
	res/drawable/widget_big_add.9.png
	res/drawable/widget_big_bg.9.png
	res/drawable/widget_big_button_bottom_right_pressed.png
	res/drawable/widget_big_ease2.9.png
	res/drawable/widget_big_ease2_rec.9.png
	res/drawable/widget_big_ease3.9.png
	res/drawable/widget_big_ease3_rec.9.png
	res/drawable/widget_big_ease4.9.png
	res/drawable/widget_big_eta.9.png
	res/drawable/widget_big_flipcard.9.png
	res/drawable/widget_big_help.9.png
	res/drawable/widget_big_opendeck.9.png
	res/layout-land/bottom_button.xml
	res/layout-small-land/studyoptions.xml
	res/layout-small/studyoptions.xml
	res/values-ar/01-core.xml
	res/values-ar/13-newfeatures.xml
	res/values-cs/02-strings.xml
	res/values-cs/03-dialogs.xml
	res/values-cs/10-preferences.xml
	res/values-cs/11-arrays.xml
	res/values-cs/13-newfeatures.xml
	res/values-de/02-strings.xml
	res/values-de/03-dialogs.xml
	res/values-de/07-cardbrowser.xml
	res/values-de/10-preferences.xml
	res/values-de/12-tutorial.xml
	res/values-el/02-strings.xml
	res/values-el/03-dialogs.xml
	res/values-el/10-preferences.xml
	res/values-el/11-arrays.xml
	res/values-el/12-tutorial.xml
	res/values-el/13-newfeatures.xml
	res/values-es/02-strings.xml
	res/values-es/03-dialogs.xml
	res/values-es/10-preferences.xml
	res/values-es/11-arrays.xml
	res/values-es/13-newfeatures.xml
	res/values-et/09-backup.xml
	res/values-et/12-tutorial.xml
	res/values-et/13-newfeatures.xml
	res/values-fr/01-core.xml
	res/values-fr/02-strings.xml
	res/values-fr/03-dialogs.xml
	res/values-fr/08-widget.xml
	res/values-fr/10-preferences.xml
	res/values-fr/11-arrays.xml
	res/values-fr/13-newfeatures.xml
	res/values-hu/02-strings.xml
	res/values-hu/03-dialogs.xml
	res/values-hu/04-network.xml
	res/values-hu/05-feedback.xml
	res/values-hu/06-statistics.xml
	res/values-hu/07-cardbrowser.xml
	res/values-hu/08-widget.xml
	res/values-hu/09-backup.xml
	res/values-hu/10-preferences.xml
	res/values-hu/11-arrays.xml
	res/values-it/01-core.xml
	res/values-it/02-strings.xml
	res/values-it/03-dialogs.xml
	res/values-it/04-network.xml
	res/values-it/06-statistics.xml
	res/values-it/07-cardbrowser.xml
	res/values-it/08-widget.xml
	res/values-it/09-backup.xml
	res/values-it/10-preferences.xml
	res/values-it/11-arrays.xml
	res/values-it/12-tutorial.xml
	res/values-it/13-newfeatures.xml
	res/values-ja/01-core.xml
	res/values-ja/02-strings.xml
	res/values-ja/03-dialogs.xml
	res/values-ja/04-network.xml
	res/values-ja/06-statistics.xml
	res/values-ja/07-cardbrowser.xml
	res/values-ja/08-widget.xml
	res/values-ja/09-backup.xml
	res/values-ja/10-preferences.xml
	res/values-ja/12-tutorial.xml
	res/values-ja/13-newfeatures.xml
	res/values-ko/01-core.xml
	res/values-nl/01-core.xml
	res/values-nl/02-strings.xml
	res/values-nl/03-dialogs.xml
	res/values-nl/08-widget.xml
	res/values-nl/09-backup.xml
	res/values-nl/10-preferences.xml
	res/values-nl/11-arrays.xml
	res/values-nl/12-tutorial.xml
	res/values-nl/13-newfeatures.xml
	res/values-no/01-core.xml
	res/values-no/02-strings.xml
	res/values-no/03-dialogs.xml
	res/values-no/08-widget.xml
	res/values-no/09-backup.xml
	res/values-no/10-preferences.xml
	res/values-pl/02-strings.xml
	res/values-pl/03-dialogs.xml
	res/values-pl/10-preferences.xml
	res/values-pl/11-arrays.xml
	res/values-pl/13-newfeatures.xml
	res/values-pt-rPT/01-core.xml
	res/values-pt-rPT/02-strings.xml
	res/values-pt-rPT/03-dialogs.xml
	res/values-pt-rPT/04-network.xml
	res/values-pt-rPT/05-feedback.xml
	res/values-pt-rPT/06-statistics.xml
	res/values-pt-rPT/07-cardbrowser.xml
	res/values-pt-rPT/08-widget.xml
	res/values-pt-rPT/10-preferences.xml
	res/values-pt-rPT/11-arrays.xml
	res/values-ru/03-dialogs.xml
	res/values-ru/10-preferences.xml
	res/values-ru/13-newfeatures.xml
	res/values-sv/02-strings.xml
	res/values-sv/03-dialogs.xml
	res/values-sv/10-preferences.xml
	res/values-sv/13-newfeatures.xml
	res/values-tr/03-dialogs.xml
	res/values-tr/10-preferences.xml
	res/values-tr/13-newfeatures.xml
	res/values-vi/01-core.xml
	res/values-vi/02-strings.xml
	res/values-vi/03-dialogs.xml
	res/values-vi/04-network.xml
	res/values-vi/05-feedback.xml
	res/values-vi/06-statistics.xml
	res/values-vi/10-preferences.xml
	res/values-vi/11-arrays.xml
	res/values-vi/13-newfeatures.xml
	res/values-zh-rCN/02-strings.xml
	res/values-zh-rCN/13-newfeatures.xml
	res/values/11-arrays.xml
	res/values/constants.xml
	src/com/ichi2/anim/ActivityTransitionAnimation.java
	src/com/ichi2/anki/About.java
	src/com/ichi2/anki/AnkiDroidApp.java
	src/com/ichi2/anki/AnkiDroidProxy.java
	src/com/ichi2/anki/CardBrowser.java
	src/com/ichi2/anki/CardEditor.java
	src/com/ichi2/anki/Deck.java
	src/com/ichi2/anki/DeckCreator.java
	src/com/ichi2/anki/DeckManager.java
	src/com/ichi2/anki/DeckPicker.java
	src/com/ichi2/anki/Feedback.java
	src/com/ichi2/anki/MetaDB.java
	src/com/ichi2/anki/MyAccount.java
	src/com/ichi2/anki/PersonalDeckPicker.java
	src/com/ichi2/anki/Preferences.java
	src/com/ichi2/anki/Reviewer.java
	src/com/ichi2/anki/SharedDeckPicker.java
	src/com/ichi2/anki/StudyOptions.java
	src/com/ichi2/anki/SyncClient.java
	src/com/ichi2/charts/ChartBuilder.java
	src/com/ichi2/libanki/Utils.java
	src/com/ichi2/themes/StyledDialog.java
	src/com/ichi2/themes/StyledProgressDialog.java
	src/com/ichi2/widget/WidgetDialog.java

+++ src/com/ichi2/anki/Feedback.java
@@ -12,16 +12,18 @@
  * this program.  If not, see <http://www.gnu.org/licenses/>.                           *
  ****************************************************************************************/
 
-package com.ichi2.anki;
+package com.ichi2.anki;import com.ichi2.anki2.R;
 
 import android.app.Application;
 import android.content.Context;
 import android.content.DialogInterface;
+import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.res.Resources;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.KeyEvent;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.WindowManager;
@@ -36,6 +38,7 @@ import android.widget.SimpleAdapter;
 import com.ichi2.anim.ActivityTransitionAnimation;
 import com.ichi2.async.Connection;
 import com.ichi2.async.Connection.Payload;
+import com.ichi2.libanki.Utils;
 import com.ichi2.themes.StyledDialog;
 import com.ichi2.themes.Themes;
 import com.tomgibara.android.veecheck.util.PrefSettings;
@@ -96,6 +99,8 @@ public class Feedback extends AnkiActivity {
     protected String mFeedbackUrl;
     protected String mErrorUrl;
 
+    private boolean mAllowFeedback;
+
     private boolean mErrorsSent = false;
 
 
@@ -132,13 +137,20 @@ public class Feedback extends AnkiActivity {
 
 
 	private void closeFeedback() {
-	        setResult(RESULT_OK);
-			finishWithAnimation(ActivityTransitionAnimation.FADE);
+	        if (getIntent().getIntExtra("request", 0) == DeckPicker.RESULT_DB_ERROR) {
+	        	setResult(DeckPicker.RESULT_DB_ERROR);
+	        } else {
+		        setResult(RESULT_OK);
+	        }
+			finish();
+            if (UIUtils.getApiLevel() > 4) {
+                ActivityTransitionAnimation.slide(Feedback.this, ActivityTransitionAnimation.LEFT);
+            }
 	}
 
 
     private void refreshInterface() {
-        if (mReportErrorMode.equals(REPORT_ASK)) {
+        if (mAllowFeedback) {
             Resources res = getResources();
             Button btnSend = (Button) findViewById(R.id.btnFeedbackSend);
             Button btnKeepLatest = (Button) findViewById(R.id.btnFeedbackKeepLatest);
@@ -203,25 +215,36 @@ public class Feedback extends AnkiActivity {
         initAllAlertDialogs();
 
         getErrorFiles();
-        if (mReportErrorMode.equals(REPORT_ALWAYS)) { // Always report
-            try {
-                String feedback = "Automatically sent";
-                Connection.sendFeedback(mSendListener, new Payload(new Object[] {
-                        mFeedbackUrl, mErrorUrl, feedback, mErrorReports, mNonce, getApplication(), true}));
-                if (mErrorReports.size() > 0) {
-                    mPostingFeedback = true;
+        Intent i = getIntent();
+        mAllowFeedback = (i.hasExtra("request") && (i.getIntExtra("request", 0) == DeckPicker.REPORT_FEEDBACK || i.getIntExtra("request", 0) == DeckPicker.RESULT_DB_ERROR)) || mReportErrorMode.equals(REPORT_ASK);
+        if (!mAllowFeedback) {
+            if (mReportErrorMode.equals(REPORT_ALWAYS)) { // Always report
+                try {
+                    String feedback = "Automatically sent";
+                    Connection.sendFeedback(mSendListener, new Payload(new Object[] {
+                            mFeedbackUrl, mErrorUrl, feedback, mErrorReports, mNonce, getApplication(), true}));
+                    if (mErrorReports.size() > 0) {
+                        mPostingFeedback = true;
+                    }
+                    if (feedback.length() > 0) {
+                        mPostingFeedback = true;
+                    }
+                } catch (Exception e) {
+                    Log.e(AnkiDroidApp.TAG, e.toString());
                 }
-                if (feedback.length() > 0) {
-                    mPostingFeedback = true;
+                finish();
+                if (UIUtils.getApiLevel() > 4) {
+                    ActivityTransitionAnimation.slide(Feedback.this, ActivityTransitionAnimation.NONE);
                 }
-            } catch (Exception e) {
-                Log.e(AnkiDroidApp.TAG, e.toString());
+                return;
+            } else if (mReportErrorMode.equals(REPORT_NEVER)) { // Never report
+                deleteFiles(false, false);
+                finish();
+                if (UIUtils.getApiLevel() > 4) {
+                    ActivityTransitionAnimation.slide(Feedback.this, ActivityTransitionAnimation.NONE);
+                }
+                return;
             }
-            closeFeedback();
-            return;
-        } else if (mReportErrorMode.equals(REPORT_NEVER)) { // Never report
-            deleteFiles(false, false);
-            closeFeedback();
         }
 
         View mainView = getLayoutInflater().inflate(R.layout.feedback, null);
@@ -308,8 +331,21 @@ public class Feedback extends AnkiActivity {
 
     }
 
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+
+        case android.R.id.home:
+        	closeFeedback();
+			return true;
+
+		default:
+			return super.onOptionsItemSelected(item);
+        }
+    }
+
     private void refreshErrorListView() {
-        if (mReportErrorMode.equals(REPORT_ASK)) {
+        if (mAllowFeedback) {
             mErrorAdapter.notifyDataSetChanged();
         }
     }
@@ -433,7 +469,7 @@ public class Feedback extends AnkiActivity {
                 }
                 refreshErrorListView();
             } else {
-                if (mReportErrorMode.equals(REPORT_ASK)) {
+                if (mAllowFeedback) {
                     if (state.equals(STATE_SUCCESSFUL)) {
                         mEtFeedbackText.setText("");
                         Themes.showThemedToast(Feedback.this, res.getString(R.string.feedback_message_sent_success), false);
